main: move http/curl code into a module
As with the sqlite code move, this allows the page-fetching to be used
reusably in other modules.
